wear: ConfirmationOverlay a11y follow-ups — (int) cast for the duration max, plus an E2E sample
Replace the out-of-class constexpr definition with an (int) cast at the
std::max call site (materializes a temporary; no ODR-use, no definition
needed). Add apps/samples/confirmationoverlay.cc: a single-button
harness that drives showOn directly (message set, duration 1000) —
verified end to end: 1000ms exact hide, 430ms fade, listener fired via
the host-root post hop, and the text-inset fraction reads 12px from the
shipped pak (454 x 2.8%).

+++ src/gui/widgetEx/wear/confirmationoverlay.cc
@@ -103,7 +103,9 @@ void ConfirmationOverlay::setUpForAccessibility() {
 // (no getSystemService).
 int ConfirmationOverlay::getDurationMillis() {
     if (AccessibilityManager::getInstance(mOverlayView->getContext()).isEnabled()) {
-        return std::max(A11Y_ANIMATION_DURATION_MS, mDurationMillis);
+        // (int) materializes a temporary for std::max's reference, avoiding
+        // the ODR-use of the constexpr member (no out-of-class definition).
+        return std::max((int) A11Y_ANIMATION_DURATION_MS, mDurationMillis);
     }
     return mDurationMillis;
 }
